Improved gulp file and new style guide
File revving option turned off by default. Can be turned on in gulpfile.js by changing
var fileRevving = true;

+++ gulpfile.js
@@ -2,28 +2,59 @@
  * PLUGINS *
  ***********/
 
-var gulp = require('gulp'),
-    sass = require('gulp-sass'),
-    rsass = require('gulp-ruby-sass'),
-    rename = require('gulp-rename'),
-    prefix = require('gulp-autoprefixer'),
-    iconfont = require('gulp-iconfont'),
-    iconfontCss = require('gulp-iconfont-css'),
-    resolver = require("gulp-resolver"),
-    rimraf = require('gulp-rimraf'),
-    shell = require('gulp-shell'),
-    cssmin = require('gulp-cssmin'),
-    cleancss = require('gulp-cleancss'),
-    gcmq = require('gulp-group-css-media-queries'),
-    uglify = require('gulp-uglify'),
-    jsonSass = require('gulp-json-sass'),
-    concat = require('gulp-concat'),
-    insert = require('gulp-insert'),
-    jshint = require('gulp-jshint'),
-    browserify = require('gulp-browserify'),
-    gulpkss = require('gulp-kss'),
-    runSequence = require('run-sequence').use(gulp);
+var gulp            = require('gulp'),
+    sass            = require('gulp-sass'),
+    rsass           = require('gulp-ruby-sass'),
+    rename          = require('gulp-rename'),
+    prefix          = require('gulp-autoprefixer'),
+    iconfont        = require('gulp-iconfont'),
+    iconfontCss     = require('gulp-iconfont-css'),
+    resolver        = require("gulp-resolver"),
+    rimraf          = require('gulp-rimraf'),
+    shell           = require('gulp-shell'),
+    cssmin          = require('gulp-cssmin'),
+    cleancss        = require('gulp-cleancss'),
+    gcmq            = require('gulp-group-css-media-queries'),
+    uglify          = require('gulp-uglify'),
+    gutil           = require('gulp-util'),
+    jsonSass        = require('gulp-json-sass'),
+    concat          = require('gulp-concat'),
+    insert          = require('gulp-insert'),
+    jshint          = require('gulp-jshint'),
+    browserify      = require('gulp-browserify'),
+    gulpkss         = require('gulp-kss'),
+    clc             = require('cli-color'),
+    runSequence     = require('run-sequence').use(gulp);
 
+
+// Colored log messages
+var error   = clc.red.bold,
+    warn    = clc.yellow,
+    notice  = clc.blue,
+    notice2 = clc.blue.bold;
+
+/**********************
+ * Icon font settings *
+ **********************/
+var fontName = 'icons',
+    fontVersion = Math.floor(Date.now()); // Timestamp in miliseconds
+
+// Apply file revving to font files? 
+// For `true`, add these lines to .htaccess files:
+// 
+//    # Turn on mod_rewrite (must be turned on in Apache httpd.conf)
+//    Options +FollowSymlinks
+//    RewriteEngine On
+//
+//    ######################
+//    # Icons file revving #
+//    ######################
+//
+//    RewriteCond %{REQUEST_FILENAME} !-f
+//    RewriteCond %{REQUEST_FILENAME} !-d
+//    RewriteRule ^(.+)\.(\d+)\.(eot|svg|woff2?|ttf)$ $1.$3 [L]
+
+var fileRevving = false;
 
 
 /*********
@@ -44,7 +75,7 @@ var paths = {
         icons   : basePaths.project + 'icons/'
     },
     styleGuide: {
-        template: basePaths.project + 'styleGuide-template/',
+        template: basePaths.project + 'styleGuideTemplate/',
         final   : basePaths.public + 'styleGuide/'
     }
 };
@@ -68,11 +99,11 @@ var appFiles = {
 
 // SASS to CSS
 gulp.task('sass', function () {
-    return gulp.src([appFiles.scss])
-		.pipe(sass({
-            errLogToConsole: true,
-            outputStyle: 'compact'
-        }))
+    return gulp.src(appFiles.scss)
+        .pipe(sass({
+        errLogToConsole: true,
+        outputStyle: 'compact'
+    }))
         .pipe(prefix('last 2 version', 'safari 5', 'ie 8', 'ie 9', 'opera 12.1', 'ios 6', 'android 4', 'Firefox >= 4'))
         .pipe(gcmq())
         .pipe(gulp.dest(paths.styles.public));
@@ -82,22 +113,26 @@ gulp.task('sass', function () {
 gulp.task('mincss', function() {
     return gulp.src(appFiles.css)
         .pipe(cleancss({
-            keepBreaks: true, 
-            keepSpecialComments: 0
-        }))    
+        keepBreaks: true, 
+        keepSpecialComments: 0
+    }))    
         .pipe(cssmin())
         .pipe(rename(function(path) {
-            path.extname = '.min.css';
-        }))
+        path.extname = '.min.css';
+    }))
         .pipe(gulp.dest(paths.styles.public));
-        
+
 });
 
-// Final build
+// SASS for DEV environment - First build icons, then build SASS
 gulp.task('libStyle', function(){ 
-    runSequence('sass', 'mincss');
+    return runSequence('cleanIcons', 'iconFontFace', 'iconGlyphs', 'iconRename', 'iconRev', 'cleanRev', 'iconSass', 'sass', 'mincss');
 });
 
+// Build SASS only, Icons must be built first
+gulp.task('libStyleLite', function(){ 
+    return runSequence('iconSass', 'sass', 'mincss');
+});
 
 
 /****************************************************
@@ -110,43 +145,38 @@ gulp.task('libStyle', function(){
 
 gulp.task('rsass', function() {
     return rsass(paths.styles.project) 
-        .on('error', function (err) { console.error('Error!', err.message); })
+        .on('error', function (err) { 
+        console.log(error('Error!', err.message)); 
+    })
         .pipe(rename(function(path) {
-            path.extname = '.ruby.css';
-        }))
-        .pipe(gulp.dest(paths.styles.public));
-});
-
-// Prefix tags and group media queries
-gulp.task('rsassBuild', function() {
-    return rsass(paths.styles.project) 
-        .on('error', function (err) { console.error('Error!', err.message); })
-        .pipe(rename(function(path) {
-            path.extname = '.ruby.css';
-        }))
+        path.extname = '.ruby.css';
+    }))
         .pipe(prefix('last 2 version', 'safari 5', 'ie 8', 'ie 9', 'opera 12.1', 'ios 6', 'android 4', 'Firefox >= 4'))
         .pipe(gcmq())
         .pipe(gulp.dest(paths.styles.public));
 });
 
+
+
 // Minimize ruby CSS
-gulp.task('rsassMin', function () {
+gulp.task('minrcss', function () {
     return gulp.src(appFiles.rcss)
         .pipe(cleancss({
-            keepBreaks: true, 
-            keepSpecialComments: 0
-        }))    
+        keepBreaks: true, 
+        keepSpecialComments: 0
+    }))    
         .pipe(cssmin())
         .pipe(rename(function(path) {
-            path.extname = '.min.css';
-        }))
+        path.extname = '.min.css';
+    }))
         .pipe(gulp.dest(paths.styles.public));
 });
 
-// Final build
+// Final ruby build
 gulp.task('rubyStyle', function(){ 
-    runSequence('rsass', 'rsassBuild', 'rsassMin');
+    return runSequence('rsass', 'minrcss');
 });
+
 
 
 
@@ -172,67 +202,74 @@ gulp.task('cleanIcons', function() {
 gulp.task('iconFontFace', function(){
     return gulp.src([appFiles.iconSvg])
         .pipe(iconfontCss({
-            fontName: fontName,
-            appendCodepoints: true,
-            path: appFiles.tplFontFace,
-            targetPath: fontName+'.scss',
-            fontPath: fontName
-        }))
+        fontName: fontName,
+        appendCodepoints: true,
+        path: appFiles.tplFontFace,
+        targetPath: fontName+'.scss',
+        fontPath: fontName
+    }))
         .pipe(iconfont({
-            fontHeight: 512,
-            normalize: true,
-            descent: 64,
-            fontName: fontName
-         }))
+        fontHeight: 512,
+        normalize: true,
+        descent: 64,
+        fontName: fontName
+    }))
         .pipe(gulp.dest(paths.fonts.project));
 });
 gulp.task('iconGlyphs', function(){
     return gulp.src([appFiles.iconSvg])
         .pipe(iconfontCss({
-            fontName: fontName,
-            appendCodepoints: true,
-            path: appFiles.tplFontGlyphs,
-            targetPath: fontName+'-glyphs.scss',
-            fontPath: fontName
-        }))
+        fontName: fontName,
+        appendCodepoints: true,
+        path: appFiles.tplFontGlyphs,
+        targetPath: fontName+'-glyphs.scss',
+        fontPath: fontName
+    }))
         .pipe(iconfont({
-            fontHeight: 512,
-            normalize: true,
-            descent: 64,
-            fontName: fontName
-        }))
+        fontHeight: 512,
+        normalize: true,
+        descent: 64,
+        fontName: fontName
+    }))
         .pipe(gulp.dest(paths.fonts.project));
 });
 
 // copy font guide to generated font folder
 gulp.task('copyFontGuide', function () {
-    console.log('nx: COPY Font guide from "'+appFiles.fontGuide+'" to "'+ paths.fonts.project+fontName + '"');
+    console.log(notice('nx: COPY Font guide from "'+appFiles.fontGuide+'" to "'+ paths.fonts.project+fontName + '"'));
     return gulp.src(appFiles.fontGuide)
         .pipe(gulp.dest(paths.fonts.project));
 });
 
+
 // 3. Append version for font files
 gulp.task('iconRename', function () {
-    return gulp.src([paths.fonts.project+'/*.*'])
-    .pipe(rename({
-        suffix: "."+fontVersion,
-    }))
-    .pipe(gulp.dest(paths.fonts.project+'/'));
+    if(fileRevving === true) {        
+        return gulp.src([paths.fonts.project+'/*.*'])
+            .pipe(rename({
+            suffix: "."+fontVersion,
+        }))
+            .pipe(gulp.dest(paths.fonts.project+'/'));
+    }
 });
 
 // 3.1. Add revving to scss
 gulp.task('iconRev', function () {
-    return gulp.src(paths.fonts.project+'/*.scss')
-        .pipe(resolver.css({
+    if(fileRevving === true) { 
+        return gulp.src(paths.fonts.project+'/*.scss')
+            .pipe(resolver.css({
             assetsDir: paths.fonts.project
         }))
-        .pipe(gulp.dest(paths.fonts.project))     
+            .pipe(gulp.dest(paths.fonts.project))  
+    }
 });
 
 // 3.2. Delete unneeded rev files
 gulp.task('cleanRev', function() {
-    return gulp.src([paths.fonts.project+'/*.*.*'], { read: false }) // much faster
-        .pipe(rimraf());
+    if(fileRevving === true) { 
+        return gulp.src([paths.fonts.project+'/*.*.*'], { read: false }) // much faster
+            .pipe(rimraf());
+    }
 });
 
 // 4. Create CSS from SCSS
@@ -244,14 +281,14 @@ gulp.task('iconSass', function () {
 
 // 5. Copy finished font to public folder
 gulp.task('copyFont', function () {
-    console.log('nx: COPY fonts from "'+appFiles.fontFiles+'" to "'+ paths.fonts.public + '"');
+    console.log(notice('nx: COPY fonts from "'+appFiles.fontFiles+'" to "'+ paths.fonts.public + '"'));
     return gulp.src(appFiles.fontFiles)
         .pipe(gulp.dest(paths.fonts.public));
 });
 
 // Delete unneeded files 
 gulp.task('cleanFontFolder', function() {
-    console.log('nx: CLEANING up Font folder');
+    console.log(notice('nx: CLEANING up Font folder'));
     return gulp.src(paths.fonts.public + '_iconFontTemplate/', { read: false }) // much faster
         .pipe(rimraf());
 });
@@ -269,7 +306,7 @@ gulp.task('buildFont', function(){
 
 // Delete old styleGuide
 gulp.task('cleanGuide', function() {
-    console.log('nx: DELETING old styleGuide');
+    console.log(notice('nx: DELETING old styleGuide'));
     return gulp.src(paths.styleGuide.final, { read: false }) // much faster
         .pipe(rimraf());
 });
@@ -277,19 +314,25 @@ gulp.task('cleanGuide', function() {
 // Create new styleGuide
 gulp.task('kss-shell', shell.task([
     // FAQ - https://github.com/kss-node/kss-node
+    'kss-node css/ styleGuide/ --template styleGuideTemplate/ --m ../public/css/style.css'
+]));
+
+gulp.task('kss-shell', shell.task([
+    // FAQ - https://github.com/kss-node/kss-node
     'kss-node '+paths.fonts.project+' '+paths.styleGuide.final +' --template '+paths.styleGuide.template+' --m *.css'
+    //    'kss-node css/ styleGuide/ --template styleGuideTemplate/ --m main.css'
 ]));
 
 // Copy assets to styleGuide folder
 gulp.task('kss-assets', function () {
-    console.log('nx: COPY fonts from "'+appFiles.fontFiles+'" to "'+ paths.styleGuide.final + 'fonts/"');
+    console.log(notice('nx: COPY fonts from "'+appFiles.fontFiles+'" to "'+ paths.styleGuide.final + 'fonts/"'));
     return gulp.src(appFiles.fontFiles)
         .pipe(gulp.dest(paths.styleGuide.final + 'fonts/'));
 });
 
 // Delete unneeded files 
 gulp.task('cleanGuide2', function() {
-    console.log('nx: CLEANING up StyleGuide');
+    console.log(notice('nx: CLEANING up StyleGuide'));
     return gulp.src(paths.styleGuide.final + 'fonts/_iconFontTemplate/', { read: false }) // much faster
         .pipe(rimraf());
 });
@@ -302,6 +345,7 @@ gulp.task('guide', function(){
 gulp.task('buildGuide', function(){ 
     runSequence('cleanGuide', 'kss-shell', 'kss-assets', 'cleanGuide2');
 });
+
 
 
 /**************
